Edit module over leerplannen heen

+++ app/Http/Livewire/Leerplan.php
@@ -14,6 +14,7 @@ class Leerplan extends _MyComponent
     public Uitvoer $uitvoer;
     public $versie_id;
     public $uitvoeren = [];
+    public $vaknaam;
 
     protected $className = \App\Models\ModuleVersie::class;
     protected $rules = [
@@ -41,29 +42,53 @@ class Leerplan extends _MyComponent
     public function setVersieItem($module_id, $vak_id)
     {
         $vak = VakInUitvoer::find($vak_id);
+        $this->vaknaam = $vak->parent->naam;
         $this->item = $vak->modules()->find($module_id);
         $this->versie_id = $this->item->id;
     }
 
+    public function editModulePreview()
+    {
+        $this->emit('confirm');
+        $this->emit('editModulePreview');
+    }
+
     public function editModule()
     {
-        if($this->versie_id == $this->item->id)
+        $vak_id = VakInUitvoer::find($this->item->pivot['vak_in_uitvoer_id'])->vak_id;
+        foreach($this->uitvoeren as $uitvoer_id)
         {
-            $this->item->vakken()->updateExistingPivot($this->item->pivot['vak_in_uitvoer_id'], [
-                'week_start' => $this->item->pivot['week_start'],
-                'week_eind' => $this->item->pivot['week_eind'],
-            ]);
-    
+            if($uitvoer_id)
+            {
+                $vak = VakInUitvoer::where('vak_id', $vak_id)->where('uitvoer_id', $uitvoer_id)->first();
+                if($vak)
+                {
+                    $module = $vak->modules()->where('module_id', $this->item->parent->id)->first();
+                    $vak->modules()->detach($module->id);
+                    $vak->modules()->attach($this->versie_id, [
+                        'week_start' => $this->item->pivot['week_start'],
+                        'week_eind' => $this->item->pivot['week_eind'],
+                    ]);
+                }
+            }
         }
-        else
-        {
-            $vak = VakInUitvoer::find($this->item->pivot['vak_in_uitvoer_id']);
-            $vak->modules()->detach($this->item->id);
-            $vak->modules()->attach($this->versie_id, [
-                'week_start' => $this->item->pivot['week_start'],
-                'week_eind' => $this->item->pivot['week_eind'],
-            ]);
-        }
+
+        // if($this->versie_id == $this->item->id)
+        // {
+        //     $this->item->vakken()->updateExistingPivot($this->item->pivot['vak_in_uitvoer_id'], [
+        //         'week_start' => $this->item->pivot['week_start'],
+        //         'week_eind' => $this->item->pivot['week_eind'],
+        //     ]);
+        // }
+        // else
+        // {
+        //     $vak = VakInUitvoer::find($this->item->pivot['vak_in_uitvoer_id']);
+        //     $vak->modules()->detach($this->item->id);
+        //     $vak->modules()->attach($this->versie_id, [
+        //         'week_start' => $this->item->pivot['week_start'],
+        //         'week_eind' => $this->item->pivot['week_eind'],
+        //     ]);
+        // }
 
         $this->endModal();
     }
@@ -82,7 +107,10 @@ class Leerplan extends _MyComponent
             if($uitvoer_id)
             {
                 $vak = VakInUitvoer::where('vak_id', $vak_id)->where('uitvoer_id', $uitvoer_id)->first();
-                $vak->modules()->detach($this->item->id);
+                if($vak)
+                {
+                    $vak->modules()->detach($this->item->id);
+                }
             }
         }
         $this->endModal();
